fix some ui issues

+++ client/src/components/UI/Buttons/TransDeleteButton.jsx
@@ -4,14 +4,18 @@ import Button from 'react-bootstrap/esm/Button'
 
 export default function TransDeleteButton({transaction, getTransactions, deleteTransaction}){
   const handleDeleteTransaction = () => {
-    deleteTransaction(transaction.id);
-
-    //refresh data
-    getTransactions();    
-  }
+    const confirmed = window.confirm("Are you sure you want to delete this transaction?");
+    if (confirmed) {
+      deleteTransaction(transaction.id);
+  
+      // Refresh data
+      getTransactions();
+    }
+  };
+  
   return (
-    <Button className='btn btn-danger' onClick={()=> handleDeleteTransaction()}>
-      <FontAwesomeIcon icon={faTrash} style={{color: "#F26B53"}}/>
+    <Button className='btn btn-danger' onClick={handleDeleteTransaction}>
+      <FontAwesomeIcon icon={faTrash} style={{ color: "#F26B53" }} />
     </Button>
-  )
+  );
 }
